requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
asttokens==3.0.0
black==24.10.0
click==8.1.7
colorama==0.4.6
decorator==5.1.1
Django==5.1.3
django-extensions==3.2.3
djangorestframework==3.15.2
executing==2.1.0
ipython==8.30.0
jedi==0.19.2
matplotlib-inline==0.1.7
mypy-extensions==1.0.0
packaging==24.2
parso==0.8.4
pathspec==0.12.1
platformdirs==4.3.6
prompt_toolkit==3.0.48
pure_eval==0.2.3
pydantic==2.10.2
pydantic_core==2.27.1
Pygments==2.18.0
pyparsing==3.2.0
rdflib==7.1.1
SPARQLWrapper==2.0.0
sqlparse==0.5.2
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
wcwidth==0.2.13
